requirements.txt
anyio==3.6.1
backports.cached-property==1.0.1
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
ecdsa==0.17.0
fastapi==0.78.0
graphql-core==3.2.1
idna==3.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
strawberry-graphql==0.114.0
typing-extensions==4.2.0
